requirements.txt
absl-py==0.9.0
affine==2.3.0
apache-beam==2.23.0
astunparse==1.6.3
attrs==19.3.0
avro-python3==1.9.2.1
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.6.20
cffi @ file:///tmp/build/80754af9/cffi_1596809843656/work
cftime==1.2.1
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
crcmod==1.7
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.2
dill==0.3.1.1
docopt==0.6.2
fastavro==0.23.6
Fiona==1.8.13.post1
fnn @ file:///home/n/Documents/Research/fnn
future==0.18.2
gast==0.3.3
geopandas==0.8.1
google-api-core==1.22.0
google-auth @ file:///tmp/build/80754af9/google-auth_1596863485713/work
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-core==1.4.1
google-cloud-pubsub==1.7.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-pasta==0.2.0
google-resumable-media==0.7.1
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.31.0
h5py @ file:///tmp/build/80754af9/h5py_1593454122442/work
hdfs==2.5.8
httplib2==0.17.4
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
iniconfig==1.0.1
joblib==0.16.0
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.1.1
matplotlib==3.3.0
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==2.0.0
more-itertools==8.4.0
munch==2.5.0
netCDF4==1.5.4
networkx==2.4
numpy==1.19.1
oauth2client==3.0.0
oauthlib==3.1.0
opt-einsum==3.1.0
packaging==20.4
pandas==1.1.0
pbr==5.4.5
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.12.3
py==1.9.0
pyarrow==0.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydot==1.4.1
PyJWT==1.7.1
pymongo==3.11.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyproj==2.6.1.post1
PySocks==1.7.1
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
rasterio==1.1.5
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rioxarray==0.0.31
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Shapely==1.7.0
six==1.15.0
sklearn==0.0
snuggs==1.4.7
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.7.24
toml==0.10.1
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
xarray==0.15.0
